SnackTime-runes-feature/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.8.1=py37_0
astor=0.8.0=py37_0
blas=1.0=mkl
ca-certificates=2019.11.27=0
certifi=2019.11.28=py37_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
gast=0.2.2=py37_0
google-pasta=0.1.8=py_0
grpcio=1.16.1=py37h351948d_1
h5py=2.9.0=py37h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
intel-openmp=2019.4=245
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
libprotobuf=3.11.2=h7bd577a_0
markdown=3.1.1=py37_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
mouseinfo=0.1.2=pypi_0
numpy=1.17.4=py37h4320e6b_0
numpy-base=1.17.4=py37hc3f5095_0
openssl=1.1.1d=he774522_3
opt_einsum=3.1.0=py_0
pillow=6.2.1=pypi_0
pip=19.3.1=py37_0
protobuf=3.11.2=py37h33f27b4_0
pyautogui=0.9.48=pypi_0
pygetwindow=0.0.8=pypi_0
pymsgbox=1.0.7=pypi_0
pyperclip=1.7.0=pypi_0
pyreadline=2.1=py37_1
pyrect=0.1.4=pypi_0
pyscreeze=0.1.25=pypi_0
python=3.7.5=h8c8aaf0_0
pytweening=1.0.3=pypi_0
scipy=1.3.2=py37h29ff71c_0
setuptools=42.0.2=py37_0
six=1.13.0=py37_0
sqlite=3.30.1=he774522_0
tensorboard=2.0.0=pyhb38c66f_1
tensorflow=2.0.0=gpu_py37h57d29ca_0
tensorflow-base=2.0.0=gpu_py37h390e234_0
tensorflow-estimator=2.0.0=pyh2649769_0
tensorflow-gpu=2.0.0=h0d30ee6_0
termcolor=1.1.0=py37_1
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
werkzeug=0.16.0=py_0
wheel=0.33.6=py37_0
wincertstore=0.2=py37_0
wrapt=1.11.2=py37he774522_0
zlib=1.2.11=h62dcd97_3

requests~=2.23.0
lcu_driver~=2.1.0